+++ src/Lit.Test/Expect.fs
@@ -17,27 +17,31 @@ open ExpectUtil
 // don't inherit JS Error. So we create our own binding here.
 [<Global("Error")>]
 type JsError(message: string) =
-    class end
+    member _.stack with get(): string = jsNative and set(v: string): unit = jsNative
 
 [<AttachMembers>]
-type AssertionError<'T>(assertion: string, ?description: string, ?actual: 'T, ?expected: 'T, ?brief: bool) =
+type AssertionError<'T>(assertion: string, ?description: string, ?actual: 'T, ?expected: 'T, ?brief: bool, ?extra: string) =
     inherit JsError(
         let brief = defaultArg brief false
         [
             "Expected " |> Some
             description |> Option.map (fun v -> $"'{v}' ")
             if not brief then actual |> Option.map (fun v -> $"{quote v} ")
-            $"to {assertion} " |> Some
-            if not brief then expected |> Option.map (fun v -> $"{quote v} ")
+            $"to {assertion}" |> Some
+            if not brief then expected |> Option.map (fun v -> $" {quote v} ")
+            extra |> Option.map (fun v -> $". {v}")
         ] |> List.choose id |> String.concat ""
     )
+    // Hide stack for cleaner reports as the test name is already shown
+    // If we use an empty string the test runner shows the message twice (?)
+    do base.stack <- "<stack hidden>"
     // Test runner requires these properties to be settable, not sure why
     member val actual = actual with get, set
     member val expected = expected with get, set
 
 type AssertionError =
-    static member Throw(assertion: string, ?description, ?actual: 'T, ?expected: 'T, ?brief: bool) =
-        AssertionError(assertion, ?description=description, ?actual=actual, ?expected=expected, ?brief=brief) |> throw
+    static member Throw(assertion: string, ?description, ?actual: 'T, ?expected: 'T, ?brief: bool, ?extra) =
+        AssertionError(assertion, ?description=description, ?actual=actual, ?expected=expected, ?brief=brief, ?extra=extra) |> throw
 
 // TODO: String and collection assertions
 [<RequireQualifiedAccess>]
@@ -74,13 +78,21 @@ module Expect =
         if not(lowerBound < actual && actual < upperBound) then
             AssertionError.Throw($"be between exclusive {lowerBound} and {upperBound}", actual=actual)
 
-    let isTrue (msg: string) (condition: 'T -> bool) (actual: 'T) =
-        if not(condition actual) then
+    let isTrue (msg: string) (condition: 'T -> bool) (argument: 'T) =
+        if not(condition argument) then
             AssertionError.Throw("be true", description=msg)
 
-    let isFalse (msg: string) (condition: 'T -> bool) (actual: 'T) =
-        if condition actual then
+    let isFalse (msg: string) (condition: 'T -> bool) (argument: 'T) =
+        if condition argument then
             AssertionError.Throw("be false", description=msg)
+
+    let map (f: 'T1 -> 'T2) (x: 'T1): 'T2 =
+        f x
+
+    let some (msg: string) (x: 'T option): 'T =
+        match x with
+        | Some x -> x
+        | None -> AssertionError.Throw("be some", description=msg)
 
     let find (msg: string) (condition: 'T -> bool) (items: 'T seq) =
         items
@@ -89,11 +101,28 @@ module Expect =
             | Some x -> x
             | None -> AssertionError.Throw("be found", description=msg)
 
-    let error (msg: string) (f: 'T -> 'Result) (actual: 'T) =
+    let errorAnd (msg: string) (f: 'T -> 'Result) (argument: 'T): exn =
+        let e =
+            try
+                let _ = f argument
+                None
+            with e ->
+                Some e
+        match e with
+        | Some e -> e
+        | None -> AssertionError.Throw("fail", description=msg)
+
+    let error (msg: string) (f: 'T -> 'Result) (argument: 'T): unit =
+        errorAnd msg f argument |> ignore
+
+    let successAnd (msg: string) (f: 'T -> 'Result) (argument: 'T): 'Result =
         try
-            let _ = f actual
-            AssertionError(msg, actual=actual) |> throw
-        with e -> e
+            f argument
+        with e ->
+            AssertionError.Throw("succeed", description=msg, extra="Error: " + e.Message)
+
+    let success (msg: string) (f: 'T -> 'Result) (argument: 'T): unit =
+        successAnd msg f argument |> ignore
 
     let beforeTimeout (msg: string) (ms: int) (pr: JS.Promise<'T>): JS.Promise<'T> =
         JS.Constructors.Promise.race [|