+++ src/Lit.Test/RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+### 1.1.0-beta-001
+
+* Improve Expect.Elmish API and other helpers
+
 ### 1.0.0
 
 * Stable release