+++ src/Lit/RELEASE_NOTES.md
@@ -1,5 +1,10 @@
+### 1.1.0-beta-002
+
+* Drop constructable Style Sheets for now
+
 ### 1.1.0-beta-001
 
+* useShadowDOM setting for LitElements
 * HMR support for LitElement
 * Constructable Style Sheets